fixed auth api and fast fetch???
