+++ frontend/src/services/authService.js
@@ -19,7 +19,7 @@ export async function createSession(email, password) {
     // Convert credentials (i.e. email:password) to Base64 token
     const credentials = btoa(`${email}:${password}`);
 
-    const response = await fetch(`${API_BASE_URL}/sessions`, {
+    const response = await fetch(`${API_BASE_URL}/api/sessions`, {
         method: "POST",
         headers: {
             Authorization: `Basic ${credentials}`,
